+++ apps/documenteditor/main/app/view/Toolbar.js
@@ -494,7 +494,8 @@ define([
                         menu: new Common.UI.Menu({
                             items: [
                                 {caption: this.mniImageFromFile, value: 'file'},
-                                {caption: this.mniImageFromUrl, value: 'url'}
+                                {caption: this.mniImageFromUrl, value: 'url'},
+                                {caption: this.mniImageFromStorage, value: 'storage'}
                             ]
                         })
                     });
@@ -2012,6 +2013,7 @@ define([
                 this.listStylesAdditionalMenuItem.setVisible(mode.canEditStyles);
                 this.btnContentControls.menu.items[4].setVisible(mode.canEditContentControl);
                 this.btnContentControls.menu.items[5].setVisible(mode.canEditContentControl);
+                this.mnuInsertImage.items[2].setVisible(this.mode.fileChoiceUrl && this.mode.fileChoiceUrl.indexOf("{documentType}")>-1);
             },
 
             onSendThemeColorSchemes: function (schemas) {
@@ -2368,7 +2370,8 @@ define([
             mniEditControls: 'Settings',
             tipControls: 'Insert content control',
             mniHighlightControls: 'Highlight settings',
-            textNoHighlight: 'No highlighting'
+            textNoHighlight: 'No highlighting',
+            mniImageFromStorage: 'Image from Storage'
         }
     })(), DE.Views.Toolbar || {}));
 });