+++ apps/presentationeditor/main/app/view/Toolbar.js
@@ -1027,12 +1027,14 @@ define([
                         new Common.UI.Menu({
                             items: [
                                 {caption: me.mniImageFromFile, value: 'file'},
-                                {caption: me.mniImageFromUrl, value: 'url'}
+                                {caption: me.mniImageFromUrl, value: 'url'},
+                                {caption: me.mniImageFromStorage, value: 'storage'}
                             ]
                         }).on('item:click', function (menu, item, e) {
                             me.fireEvent('insert:image', [item.value]);
                         })
                     );
+                    btn.menu.items[2].setVisible(config.fileChoiceUrl && config.fileChoiceUrl.indexOf("{documentType}")>-1);
                 });
 
                 me.btnsInsertText.forEach(function (btn) {
@@ -1731,7 +1733,8 @@ define([
             textSurface: 'Surface',
             textShowPresenterView: 'Show presenter view',
             textTabCollaboration: 'Collaboration',
-            textTabProtect: 'Protection'
+            textTabProtect: 'Protection',
+            mniImageFromStorage: 'Image from Storage'
         }
     }()), PE.Views.Toolbar || {}));
 });