+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Save to pdfa format
 * Add rotation and flip to image and shape settings (bug #19378)
 * Save file copy to selected folder (bug #23603, bug #32790)
+* Load image from storage
 
 ### Document Editor